+++ src/cjs/Bounds.js
@@ -14,7 +14,8 @@
  * @modified 2025-04-18 Change parameter type in `Bounds.computeFromVertices` from `Vertex` to more general `XYCoords`.
  * @modified 2025-04-19 Added methods to `Bounds` class: `getNorthPoint`, `getSouthPoint`, `getEastPoint` and `getWestPoint`.
  * @modified 2025-04-26 Added static method `Bounds.computeFromBoundsSet` to calculate containing bounds for a set of bounding boxes.
- * @version  1.8.0
+ * @modified 2025-10-17 Added the methods `Bounds.getSouthWestPoint`, `getNorthWestPoint`, `getNorthEastPoint` and `getSouthEastPoint`.
+ * @version  1.9.0
  **/
 Object.defineProperty(exports, "__esModule", { value: true });
 exports.Bounds = void 0;
@@ -53,7 +54,6 @@ var Bounds = /** @class */ (function () {
     Bounds.prototype.getNorthPoint = function () {
         return new Vertex_1.Vertex(this.min.x + this.width / 2.0, this.min.y);
     };
-    ;
     /**
      * Get the center point of the south bound.
      *
@@ -65,31 +65,72 @@ var Bounds = /** @class */ (function () {
     Bounds.prototype.getSouthPoint = function () {
         return new Vertex_1.Vertex(this.min.x + this.width / 2.0, this.max.y);
     };
-    ;
     /**
-    * Get the center point of the west bound.
-    *
-    * @method getWestPoint
-    * @instance
-    * @memberof Bounds
-    * @return {Vertex} The "westhmost" centered point of this bounding box.
-    */
+     * Get the center point of the west bound.
+     *
+     * @method getWestPoint
+     * @instance
+     * @memberof Bounds
+     * @return {Vertex} The "westhmost" centered point of this bounding box.
+     */
     Bounds.prototype.getWestPoint = function () {
         return new Vertex_1.Vertex(this.min.x, this.min.y + this.height / 2.0);
     };
-    ;
     /**
-    * Get the center point of the east bound.
-    *
-    * @method getEastPoint
-    * @instance
-    * @memberof Bounds
-    * @return {Vertex} The "easthmost" centered point of this bounding box.
-    */
+     * Get the center point of the east bound.
+     *
+     * @method getEastPoint
+     * @instance
+     * @memberof Bounds
+     * @return {Vertex} The "easthmost" centered point of this bounding box.
+     */
     Bounds.prototype.getEastPoint = function () {
         return new Vertex_1.Vertex(this.max.x, this.min.y + this.height / 2.0);
     };
-    ;
+    /**
+     * Get the upper left corner point these bounds box.
+     *
+     * @method getNorthWestPoint
+     * @instance
+     * @memberof Bounds
+     * @return {Vertex} The "northwestmost" point of this bounding box.
+     */
+    Bounds.prototype.getNorthWestPoint = function () {
+        return new Vertex_1.Vertex(this.min.x, this.min.y);
+    };
+    /**
+     * Get the upper right corner point these bounds box.
+     *
+     * @method getNorthEastPoint
+     * @instance
+     * @memberof Bounds
+     * @return {Vertex} The "northeastmost" point of this bounding box.
+     */
+    Bounds.prototype.getNorthEastPoint = function () {
+        return new Vertex_1.Vertex(this.max.x, this.min.y);
+    };
+    /**
+     * Get the lower right corner point these bounds box.
+     *
+     * @method getSouthEastPoint
+     * @instance
+     * @memberof Bounds
+     * @return {Vertex} The "southeastmost" point of this bounding box.
+     */
+    Bounds.prototype.getSouthEastPoint = function () {
+        return new Vertex_1.Vertex(this.max.x, this.max.y);
+    };
+    /**
+     * Get the lower left corner point these bounds box.
+     *
+     * @method getSouthWestPoint
+     * @instance
+     * @memberof Bounds
+     * @return {Vertex} The "southwestmost" point of this bounding box.
+     */
+    Bounds.prototype.getSouthWestPoint = function () {
+        return new Vertex_1.Vertex(this.min.x, this.max.y);
+    };
     /**
      * Convert this rectangular bounding box to a polygon with four vertices.
      *
@@ -202,16 +243,16 @@ var Bounds = /** @class */ (function () {
         return new Bounds(new Vertex_1.Vertex(xMin, yMin), new Vertex_1.Vertex(xMax, yMax));
     };
     /**
-    * Compute the minimal bounding box for a given set of existing bounding boxes.
-    *
-    * An empty vertex array will return an empty bounding box located at (0,0).
-    *
-    * @static
-    * @method computeFromBoundsSet
-    * @memberof Bounds
-    * @param {Array<IBounds>} boundingBoxes - The set of existing bounding boxes to get the containing bounding box for.
-    * @return The minimal Bounds for the given bounds instances.
-    **/
+     * Compute the minimal bounding box for a given set of existing bounding boxes.
+     *
+     * An empty vertex array will return an empty bounding box located at (0,0).
+     *
+     * @static
+     * @method computeFromBoundsSet
+     * @memberof Bounds
+     * @param {Array<IBounds>} boundingBoxes - The set of existing bounding boxes to get the containing bounding box for.
+     * @return The minimal Bounds for the given bounds instances.
+     **/
     Bounds.computeFromBoundsSet = function (boundingBoxes) {
         if (boundingBoxes.length == 0) {
             return new Bounds(new Vertex_1.Vertex(0, 0), new Vertex_1.Vertex(0, 0));