+++ src/cjs/utils/datastructures/DataGrid2dArrayMatrix.js
@@ -41,7 +41,7 @@ var DataGrid2dArrayMatrix = /** @class */ (function () {
     DataGrid2dArrayMatrix.prototype.find = function (condition) {
         for (var x = 0; x < this.xSegmentCount; x++) {
             for (var y = 0; y < this.ySegmentCount; y++) {
-                if (condition(this.get(x, y), x, y)) {
+                if (condition(this.get(x, y), x, y, this)) {
                     return {
                         xIndex: x,
                         yIndex: y,