+++ changelog.md
@@ -1,5 +1,10 @@
 # Changelog
 
+- 2025.10-17
+  - Added the methods `Bounds.getSouthWestPoint`, `getNorthWestPoint`, `getNorthEastPoint` and `getSouthEastPoint`.
+- 2025-10-16
+  - Added demo `65-rounded-squares`.
+  - Adding the `utils/datastructures/DataGrid2dArrayMatrix` class for array like data matrices.
 - 2025-10-04
   - Adding `src/ts/utils/SVGPathUtils` with the `mkCircularRingSector` function.
   - Adding the demo `63-simple-ring-animation`.